+++ pcs/lib/booth/cib.py
@@ -1,0 +1,29 @@
+from typing import (
+    Optional,
+    cast,
+)
+
+from lxml.etree import _Element
+
+from pcs.lib.booth.constants import DEFAULT_INSTANCE_NAME
+
+_BOOTH_ATTRIBUTE = "booth-cfg-name"
+
+
+def get_ticket_names(
+    cib: _Element, booth_instance: Optional[str] = None
+) -> list[str]:
+    """
+    Return names of booth tickets present in CIB for the specified booth
+    instance
+
+    booth_instance -- name of the booth instance, default instance name is used
+        if name is not provided
+    """
+    return cast(
+        list[str],
+        cib.xpath(
+            f"status/tickets/ticket_state[@{_BOOTH_ATTRIBUTE}=$instance]/@id",
+            instance=booth_instance or DEFAULT_INSTANCE_NAME,
+        ),
+    )